refactor: move options to src
