+++ packages/@lwc/compiler/src/bundler/bundler.ts
@@ -21,7 +21,7 @@ import rollupTransform from '../rollup-plugins/transform';
 import rollupCompat from '../rollup-plugins/compat';
 import rollupMinify from '../rollup-plugins/minify';
 
-import { NormalizedCompileOptions, validateNormalizedOptions } from '../compiler/options';
+import { NormalizedCompileOptions, validateNormalizedCompileOptions } from '../options';
 
 import { SourceMap } from '../compiler/compiler';
 
@@ -67,7 +67,7 @@ function handleRollupWarning(diagnostics: CompilerDiagnostic[]) {
 }
 
 export async function bundle(options: NormalizedCompileOptions): Promise<BundleReport> {
-    validateNormalizedOptions(options);
+    validateNormalizedCompileOptions(options);
 
     const { outputConfig, name, namespace } = options;
