+++ packages/@lwc/compiler/src/transformers/__tests__/transform-html.spec.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import { TransformOptions } from '../../compiler/options';
+import { TransformOptions } from '../../options';
 import { transform } from '../transformer';
 
 import { pretify } from '../../__tests__/utils';
@@ -12,7 +12,6 @@ import { pretify } from '../../__tests__/utils';
 const TRANSFORMATION_OPTIONS: TransformOptions = {
     namespace: 'x',
     name: 'foo',
-    files: {},
 };
 
 it('should throw when processing an invalid HTML file', async () => {
